Move to geojson_writer mod
