+++ rust/src/writer/mod.rs
@@ -1,0 +1,3 @@
+mod geojson_writer;
+
+pub(crate) use geojson_writer::write_geojson;
